requirements.txt
Django==3.1.3
djangorestframework==3.14.0
scipy==1.9.1
setfit==0.7.0
transformers==4.30.2
